requirements.txt
schedule
pytz
requests
python-dotenv
google-genai